added BaseVerifyServerSettingsActivity class
